refac center method to utils file
