+++ src/ui.rs
@@ -16,6 +16,7 @@ pub mod colors;
 pub mod file_menu;
 pub mod string_list;
 pub mod talbe_view;
+pub mod utils;
 
 const APP_NAME: &str = " JDbrowser ";
